backend/requirements.txt
fastapi==0.115.9
uvicorn==0.40.0
PyMuPDF==1.27.1
python-docx==1.2.0
numpy==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.31.0
openai>=1.0.0
pymongo>=4.6.0